created component for transaction range selection
